Environments & agent progress ion backgammon
